% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LEAVcore-package.R
\docType{package}
\name{LEAVcore-package}
\alias{LEAVcore}
\alias{LEAVcore-package}
\title{LEAVcore: Constitution of Core Collections using Length of Encoded Attribute Values}
\description{
Construct core collections using the information measure 'Length of Encoded Attribute Values' (LEAV) using qualitative and/or quantitative trait data as described by Balakrishnan and Suresh (2001a) \url{https://indianjournals.com/article/ijpgr-14-1-006} and (2001b) \url{https://indianjournals.com/article/ijpgr-14-3-005}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/aravind-j/LEAVcore} \url{https://aravind-j.github.io/LEAVcore/}
  \item Report bugs at \url{https://github.com/aravind-j/LEAVcore/issues}
}

}
\author{
\strong{Maintainer}: J. Aravind \email{j.aravind@icar.org.in} (\href{https://orcid.org/0000-0002-4791-442X}{ORCID})

Authors:
\itemize{
  \item J. Aravind \email{j.aravind@icar.org.in} (\href{https://orcid.org/0000-0002-4791-442X}{ORCID})
  \item Suman Roy \email{suman.roy@icar.org.in} (\href{https://orcid.org/0000-0001-5612-8415}{ORCID})
  \item Anju Mahendru Singh \email{anju.singh@icar.org.in} (\href{https://orcid.org/0000-0001-6958-1630}{ORCID})
}

Other contributors:
\itemize{
  \item ICAR-NBGPR (\href{https://ror.org/00scbd467}{ROR}) (url: https://nbpgr.org.in/) [copyright holder]
}

}
\keyword{internal}
