% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inflen.qual.R
\name{inflen.qual}
\alias{inflen.qual}
\title{Compute Information Length for Qualitative Traits}
\usage{
inflen.qual(x, freq, adj = TRUE)
}
\arguments{
\item{x}{Data of a qualitative trait for accessions in a collection as a
vector of type factor.}

\item{freq}{The absolute frequencies of the descriptor states of the
qualitative trait \code{x} in the subset of all accessions.}

\item{adj}{logical. If \code{TRUE}, the proportion estimates are slightly
biased to include zero frequency descriptor states in the computation (See
\strong{Details}). Default is \code{TRUE}.}
}
\value{
A data frame with 2 columns: \describe{ \item{x}{The qualitative
  trait data} \item{inflen}{Information length computed}}
}
\description{
The function \code{inflen.qual} computes the length of information code that
can indicate the possession of a descriptor state of a qualitative trait
\insertCite{wallace_information_1968,balakrishnan_Strategies_2001,balakrishnan_Strategies_2001a,balakrishnan_Strategies_2003}{LEAVcore}
\loadmathjax.
}
\details{
For each qualitative trait/descriptor \mjseqn{d} the probability of
occurrence of a descriptor state \mjseqn{m} in the in a subset \mjseqn{t} is
estimated as

\mjsdeqn{p_{m,d,t} = \frac{n_{m,d,t} + 1}{n_{d,t} + M_{d}}}

Where, \mjseqn{n_{m,d,t}} is the number of accessions with \mjseqn{m} state
of trait \mjseqn{d} in subset \mjseqn{t}, \mjseqn{n_{d,t}} is the number of
accessions with any known state of trait \mjseqn{d} in subset \mjseqn{t},i.e.
the number of accessions in subset \mjseqn{t} and \mjseqn{M_{d}} is the
number of descriptor states of trait \mjseqn{d}.

This is a slightly biased estimate to include zero frequency descriptor
states in the computation. The actual estimate is

\mjsdeqn{p_{m,d,t} = \frac{n_{m,d,t}}{n_{d,t}}}

Now the length of the information code that can optimally indicate the
possession of descriptor state \mjseqn{m} of trait \mjseqn{d} in the subset
\mjseqn{t} is computed as

\mjsdeqn{c_{m,d,t} = -\ln p_{m,d,t}}
}
\examples{

suppressPackageStartupMessages(library(EvaluateCore))

library(EvaluateCore)

# Get data from EvaluateCore

data("cassava_EC", package = "EvaluateCore")

# Data of 'Colour of unexpanded apical leaves' qualitative trait
CUAL <- as.factor(cassava_EC$CUAL)

# Get frequencies based on sample size
prop <-  prop.adj(CUAL, method = "sqrt")
size.prop <- 0.2
size.count <- ceiling(size.prop * length(CUAL))
CUALfreq <- round(prop * size.count)

# Compute information length
CUALinflen <- inflen.qual(x = CUAL, freq = CUALfreq, adj = TRUE)

head(CUALinflen)

}
\references{
\insertAllCited{}
}
