% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LEAV.R
\name{LEAV}
\alias{LEAV}
\title{Compute the Length of Encoded Attribute Values}
\usage{
LEAV(
  data,
  names,
  quantitative = NULL,
  qualitative = NULL,
  freq,
  adj = TRUE,
  mean,
  sd,
  e
)
}
\arguments{
\item{data}{The data as a data frame object. The data frame should possess
one row per individual and columns with the individual names and multiple
trait/character data.}

\item{names}{Name of column with the individual names as a character string.}

\item{quantitative}{Name of columns with the quantitative traits as a
character vector.}

\item{qualitative}{Name of columns with the qualitative traits as a character
vector.}

\item{freq}{A named list with the target absolute frequencies of the
descriptor states for each qualitative trait specified in
\code{qualitative}. The list names should be same as \code{qualitative}.}

\item{adj}{logical. If \code{TRUE}, the proportion estimates are slightly
biased to include zero frequency descriptor states in the computation (See
\strong{Details}). Default is \code{TRUE}.}

\item{mean}{A named numeric vector of target means for each quantitative
trait specified in \code{quantitative}. The list names should be same as
\code{quantitative}.}

\item{sd}{A named numeric vector of target standard deviation for each
quantitative trait specified in \code{quantitative}. The list names should
be same as \code{quantitative}.}

\item{e}{A named numeric vector of least count of measurement for each
quantitative trait specified in \code{quantitative}. The list names should
be same as \code{quantitative}.}
}
\value{
A data frame with one row per accession in \code{data} and the
  following columns:
  \describe{
    \item{\code{names}}{Accession identifiers, as specified by the
      \code{names} argument.}
    \item{\code{lt}}{The log-ratio message length term,
      \mjseqn{\log(N / n)}, where \mjseqn{N} is the total number of
      accessions in \code{data} and \mjseqn{n} is the sum of frequencies in
      \code{freq}.}
    \item{\code{<qualitative traits>}}{One column per trait specified in
      \code{qualitative}, giving the information length
      \mjseqn{-\log(p_k)} for the level \mjseqn{k} of that trait observed
      for each accession.}
    \item{\code{<quantitative traits>}}{One column per trait specified in
      \code{quantitative}, giving the Gaussian information length
      \mjseqn{\log(\sigma / c \varepsilon) + (x - \mu)^2 / 2\sigma^2}
      for each accession, where \mjseqn{c = 1/\sqrt{2\pi}}.}
    \item{\code{LEAV}}{The total information length for each accession,
      equal to the row sum of \code{lt} and all trait information length
      columns.}
  }
}
\description{
For accessions in a collection compute the Length of Encoded Attribute Values
(LEAV) information measure from qualitative and quantitative trait data
\insertCite{wallace_information_1968,balakrishnan_Strategies_2001,balakrishnan_Strategies_2001a,balakrishnan_Strategies_2003}{LEAVcore}.
\loadmathjax
}
\details{
For each accession \mjseqn{s} in the collection, the message length
\mjseqn{F_{s}} to optimally encode all the \mjseqn{d} traits/descriptors is
computed as follows using the joint density distribution of the whole
collection.

\mjsdeqn{F_{s} = l_{t} + \sum_{i=1}^{p} c_{m_{s},d_{i},t} +
\sum_{j=1}^{q} c_{x_{s},d_{j},t}}

Here, the first expression \mjseqn{l_{t}} is the message length for the
subset \mjseqn{t} to which an accession belongs when there are \mjseqn{N}
accessions in the whole collection and \mjseqn{n_{t}} accessions in the
subset \mjseqn{t}.

\mjsdeqn{l_{t} = l_{t} = - \ln{\left ( \frac{N}{n_{t}} \right )}}

Similarly \mjseqn{\sum_{i=1}^{p} c_{m_{s},d_{i},t}} is sum of the optimum
message length for \mjseqn{p} qualitative traits, \mjseqn{\sum_{j=1}^{q}
c_{x_{s},d_{j},t}} is sum of the optimum optimum message length for
\mjseqn{q} quantitative traits. See \code{\link[LEAVcore]{inflen.qual}} and
\code{\link[LEAVcore]{inflen.quant}} for more details.
}
\examples{
suppressPackageStartupMessages(library(EvaluateCore))

# Get data from EvaluateCore
data("cassava_EC", package = "EvaluateCore")

cassava_EC <- cbind(genotypes = rownames(cassava_EC), cassava_EC)


quant <- c("NMSR", "TTRN", "TFWSR", "TTRW", "TFWSS", "TTSW", "TTPW", "AVPW",
                  "ARSR", "SRDM")
qual <- c("CUAL", "LNGS", "PTLC", "DSTA", "LFRT", "LBTEF", "CBTR", "NMLB",
                 "ANGB", "CUAL9M", "LVC9M", "TNPR9M", "PL9M", "STRP", "STRC",
                 "PSTR")

cassava_EC[, qual] <- lapply(cassava_EC[, qual], as.factor)

size <- 0.2

freq_list <- lapply(qual, function(x) {
  prop <-  prop.adj(cassava_EC[, x], method = "sqrt")
  size.count <- ceiling(size * length(x))
  round_preserve_sum(prop * size.count)
})
names(freq_list) <- qual

mean_vec <- sapply(cassava_EC[, quant],
                   function(x) {
                     floor(mean(x))
                   })
names(mean_vec) <- quant

sd_vec <- sapply(cassava_EC[quant],
             function(x) {
               round(sd(x), 1)
             })
names(sd_vec) <- quant

e_vec <- rep(1, length(quant))
names(e_vec) <- quant


# Compute LEAV
LEAV_cassava <- LEAV(data = cassava_EC, names = "genotypes",
                     quantitative = quant, qualitative = qual,
                     freq = freq_list, adj = TRUE,
                     mean = mean_vec, sd = sd_vec, e = e_vec)

head(LEAV_cassava)

}
\references{
\insertAllCited{}
}
\seealso{
\code{\link[LEAVcore]{inflen.qual}},
  \code{\link[LEAVcore]{inflen.quant}}
}
