% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LEAVcore_functions.R
\name{LEAVcore_functions}
\alias{LEAVcore_functions}
\alias{LEAVcore1}
\alias{LEAVcore2}
\alias{LEAVcore3}
\title{Generate Core collections with Length of Encoded Attribute Values}
\usage{
LEAVcore1(
  data,
  names,
  quantitative = NULL,
  qualitative = NULL,
  sample.count,
  prop.adj = c("none", "log", "sqrt"),
  e
)

LEAVcore2(
  data,
  names,
  quantitative = NULL,
  qualitative = NULL,
  sample.count,
  prop.adj = c("none", "log", "sqrt"),
  e
)

LEAVcore3(
  data,
  names,
  quantitative = NULL,
  qualitative = NULL,
  sample.count,
  prop.adj = c("none", "log", "sqrt"),
  e
)
}
\arguments{
\item{data}{The data as a data frame object. The data frame should possess
one row per individual and columns with the individual names and multiple
trait/character data.}

\item{names}{Name of column with the individual names as a character string.}

\item{quantitative}{Name of columns with the quantitative traits as a
character vector.}

\item{qualitative}{Name of columns with the qualitative traits as a character
vector.}
}
\description{
Based on Length of Encoded Attribute Values (LEAV)
\insertCite{wallace_information_1968,balakrishnan_Strategies_2001,balakrishnan_Strategies_2001a,balakrishnan_Strategies_2003}{LEAVcore}
estimated from qualitative and/or quantitative trait data, core collections
can be generated by the three following methods. \describe{
\item{Method I}{Classification based on pre-determined diversity represented
by LEAV estimates implemented in \code{LEAVcore1}.}
\item{Method II}{Purposive selection of  accessions with highest rank of
LEAV estimates implemented in \code{LEAVcore2}.}
\item{Method III}{Stratified sampling of accessions from diversity
groups/strata computed from LEAV estimates partially implemented in
\code{LEAVcore3}.}
}.
\loadmathjax
}
