% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LEAVcore_functions.R
\name{LEAVcore_functions}
\alias{LEAVcore_functions}
\alias{LEAVcore1}
\alias{LEAVcore2}
\alias{LEAVcore3}
\title{Generate Core collections with Length of Encoded Attribute Values}
\usage{
LEAVcore1(
  data,
  names,
  quantitative = NULL,
  qualitative = NULL,
  size,
  prop.adj = c("none", "log", "sqrt"),
  e
)

LEAVcore2(
  data,
  names,
  quantitative = NULL,
  qualitative = NULL,
  size,
  prop.adj = c("none", "log", "sqrt"),
  e
)

LEAVcore3(
  data,
  names,
  quantitative = NULL,
  qualitative = NULL,
  size,
  prop.adj = c("none", "log", "sqrt"),
  e
)
}
\arguments{
\item{data}{The data as a data frame object. The data frame should possess
one row per individual and columns with the individual names and multiple
trait/character data.}

\item{names}{Name of column with the individual names as a character string.}

\item{quantitative}{Name of columns with the quantitative traits as a
character vector.}

\item{qualitative}{Name of columns with the qualitative traits as a character
vector.}

\item{size}{The desired core set size proportion.}

\item{prop.adj}{The method for relative frequency transformation for
qualitative traits. Either \code{"none"} for no transformation or
\code{"log"} for log-frequency transformation or \code{"sqrt"} for square
root-proportion transformation (see \code{\link[LEAVcore]{prop.adj}}).}

\item{e}{A named numeric vector of least count of measurement for each
quantitative trait specified in \code{quantitative}. The list names should
be same as \code{quantitative}.}
}
\description{
Based on Length of Encoded Attribute Values (LEAV)
\insertCite{wallace_information_1968,balakrishnan_Strategies_2001,balakrishnan_Strategies_2001a,balakrishnan_Strategies_2003}{LEAVcore}
estimated from qualitative and/or quantitative trait data, core collections
can be generated by the three following methods. \describe{
\item{Method I}{Classification based on pre-determined diversity represented
by LEAV estimates implemented in \code{LEAVcore1}.}
\item{Method II}{Purposive selection of  accessions with highest rank of
LEAV estimates implemented in \code{LEAVcore2}.}
\item{Method III}{Stratified sampling of accessions from diversity
groups/strata computed from LEAV estimates partially implemented in
\code{LEAVcore3}.}
}.
\loadmathjax
}
\details{
\insertCite{balakrishnan_Strategies_2001,balakrishnan_Strategies_2001a;textual}{LEAVcore}
describe three different methods of constructing core collections from
estimates of Length of Encoded Attribute Values.

 \subsection{Method I: Classification based on pre-determined diversity represented
 by LEAV estimates}{

 }

 \subsection{Method II: Purposive selection of  accessions with highest rank of
LEAV estimates}{

 }

 \subsection{Method III: Stratified sampling of accessions from diversity
groups/strata computed from LEAV estimates}{

 }
}
