% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LEAVcore_functions.R
\name{LEAVcore_functions}
\alias{LEAVcore_functions}
\alias{LEAVcore1}
\alias{LEAVcore2}
\title{Generate Core collections with Length of Encoded Attribute Values}
\usage{
LEAVcore1(
  data,
  names,
  quantitative = NULL,
  qualitative = NULL,
  sample.count,
  prop.adj = c("none", "log", "sqrt"),
  e
)

LEAVcore2(
  data,
  names,
  quantitative = NULL,
  qualitative = NULL,
  sample.count,
  prop.adj = c("none", "log", "sqrt"),
  e
)
}
\arguments{
\item{data}{The data as a data frame object. The data frame should possess
one row per individual and columns with the individual names and multiple
trait/character data.}

\item{names}{Name of column with the individual names as a character string.}

\item{quantitative}{Name of columns with the quantitative traits as a
character vector.}

\item{qualitative}{Name of columns with the qualitative traits as a character
vector.}
}
\description{
Generate a core collection with Length of Encoded Attribute Values (LEAV)
\insertCite{wallace_information_1968,balakrishnan_Strategies_2001,balakrishnan_Strategies_2001a,balakrishnan_Strategies_2003}{LEAVcore}
estimated from qualitative and/or quantitative trait data
\itemize{\item{LEAVcore1}{} \item{}{LEAVcore2}}. \loadmathjax
}
