% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LEAVcore_functions.R
\name{LEAVcore_functions}
\alias{LEAVcore_functions}
\alias{LEAVcore1}
\alias{LEAVcore2}
\alias{LEAVcore3}
\title{Generate Core collections with Length of Encoded Attribute Values}
\usage{
LEAVcore1(
  data,
  names,
  quantitative = NULL,
  qualitative = NULL,
  sample.count,
  prop.adj = c("none", "log", "sqrt"),
  e
)

LEAVcore2(
  data,
  names,
  quantitative = NULL,
  qualitative = NULL,
  sample.count,
  prop.adj = c("none", "log", "sqrt"),
  e
)

LEAVcore3(
  data,
  names,
  quantitative = NULL,
  qualitative = NULL,
  sample.count,
  prop.adj = c("none", "log", "sqrt"),
  e
)
}
\arguments{
\item{e}{}
}
\description{

}
