% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/round_preserve_sum.R
\name{round_preserve_sum}
\alias{round_preserve_sum}
\title{Round Numeric Values While Preserving a Target Sum}
\usage{
round_preserve_sum(x, target = round(sum(x)))
}
\arguments{
\item{x}{A numeric vector to round.}

\item{target}{A numeric scalar giving the desired sum of the rounded values.
Defaults to \code{round(sum(x))}.}
}
\value{
An numeric vector of the same length as \code{x}, where the elements
  sum to \code{target}.
}
\description{
Applies the Hamilton (largest remainder or Hare-Niemeyer or Vinton) rounding
method \insertCite{balinski_fair_2001}{LEAVcore} to a numeric vector so that
the rounded values sum to a specified target.
}
\details{
Values are first rounded down using \code{floor()}, and the remaining deficit
is allocated to elements with the largest fractional parts.
}
\examples{
round_preserve_sum(c(1.2, 2.7, 3.5))

round_preserve_sum(c(10.4, 10.4, 10.2), target = 32)

}
\references{
\insertAllCited{}
}
